// api/src/core/repositories/UserRepository/dto/updateUserDto.d.ts
export declare class UpdateUserDto {
    readonly name?: string | undefined;
    readonly email?: string | undefined;
    readonly password?: string | undefined;
    constructor(name?: string | undefined, email?: string | undefined, password?: string | undefined);
}
//# sourceMappingURL=updateUserDto.d.ts.map